+++ src/Application.hpp
@@ -27,6 +27,7 @@ class Fonts;
 class Toasts;
 class ChatterinoBadges;
 class DankerinoBadges;
+class SeventvBadges;
 class FfzBadges;
 
 class Application
@@ -60,6 +61,7 @@ public:
     TwitchIrcServer *const twitch2{};
     ChatterinoBadges *const chatterinoBadges{};
     DankerinoBadges *const dankerinoBadges{};
+    SeventvBadges *const seventvBadges{};
     FfzBadges *const ffzBadges{};
 
     /*[[deprecated]]*/ Logging *const logging{};