+++ src/messages/MessageElement.hpp
@@ -51,6 +51,7 @@ enum class MessageElementFlag : int64_t {
 
     SeventvEmoteImage = (1LL << 34),
     SeventvEmoteText = (1LL << 35),
+
     SeventvEmote = SeventvEmoteImage | SeventvEmoteText,
 
     EmoteImages =
@@ -97,7 +98,14 @@ enum class MessageElementFlag : int64_t {
     // - Chatterino gnome badge
     BadgeChatterino = (1LL << 18),
 
-    // Slot 7: FrankerFaceZ
+    // Slot 7: Seventv
+    // - Seventv Admin
+    // - Seventv Dungeon Mistress
+    // - Seventv Moderator
+    // - Seventv Subscriber
+    BadgeSeventv = (1LL << 40),
+
+    // Slot 8: FrankerFaceZ
     // - FFZ developer badge
     // - FFZ bot badge
     // - FFZ donator badge
@@ -109,7 +117,7 @@ enum class MessageElementFlag : int64_t {
 
     Badges = BadgeGlobalAuthority | BadgePredictions | BadgeChannelAuthority |
              BadgeSubscription | BadgeVanity | BadgeChatterino | BadgeFfz |
-             BadgeDankerino,
+             BadgeDankerino | BadgeSeventv,
 
     ChannelName = (1LL << 20),