+++ src/providers/twitch/TwitchMessageBuilder.hpp
@@ -87,6 +87,7 @@ private:
     void appendTwitchBadges();
     void appendChatterinoBadges();
     void appendDankerinoBadges();
+    void appendSeventvBadges();
     void appendFfzBadges();
     Outcome tryParseCheermote(const QString &string);